LZMR-47 modified characters service layer
