+++ Netbeans/SuperheroSightings/src/main/java/com/sg/superherosightings/dao/CharactersDao.java
@@ -30,6 +30,10 @@ public interface CharactersDao{
     public List<Characters> getAllCharacters();
     
     public List<Characters> getAssociatedCharacters(List<Sighting> sightings);
+    
+    public List<Characters> getAllHeroes();
+
+    public List<Characters> getAllVillains();
 
     
 }
